recommendation-system/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
scikit-learn==1.4.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
pytest==8.0.2 